[AIRFLOW-3765] Fix merge conflicts; Move custom SQLAlchemy types to its own file
